updated visualMedia
