+++ packages/teams-js/src/public/index.ts
@@ -78,7 +78,7 @@ export * as search from './search';
 export * as sharing from './sharing/sharing';
 export * as stageView from './stageView/stageView';
 export { version } from './version';
-export { visualMedia } from './visualMedia';
+export * as visualMedia from './visualMedia/visualMedia';
 export { webStorage } from './webStorage';
 export * as call from './call';
 export * as appInitialization from './appInitialization';
